requirements.txt
amqp==1.4.6
anyjson==0.3.3
astroid==1.3.6
beautifulsoup4==4.3.2
billiard==3.3.0.19
celery==3.1.17
Django==1.7.7
django-appconf==1.0.1
django-braces==1.4.0
django-compressor==1.4
django-crispy-forms==1.4.0
django-extensions==1.5.1
django-gravatar2==1.3.0
django-redis==3.8.3
django-su==0.4.7
djangorestframework==3.1.0
gnureadline==6.3.3
html5lib==0.999
httplib2==0.9
ipython==3.1.0
kombu==3.0.24
-e git+https://github.com/ducksboard/libsaas@dbf02ae10b3c8ef816c8cdf094ed3d80e866d7e2#egg=libsaas-master
logilab-common==0.63.2
oauth2==1.5.211
oauthlib==0.7.2
psycopg2==2.6
PyJWT==1.0.1
python-openid==2.2.5
python-slugify==1.0.2
python-social-auth==0.2.6
pytz==2015.2
redis==2.10.3
requests==2.6.0
requests-oauthlib==0.4.2
six==1.9.0
Unidecode==0.4.17
Werkzeug==0.10.4
